joi check moved to controllers
