+++ controllers/eventController.js
@@ -1,4 +1,31 @@
 const Event = require("../models/Event");
+const express = require("express");
+const Joi = require("joi");
+// Defining a Checking schema for the Event Body
+const minDate = `1-1-${new Date(Date.now()).getFullYear() - 1}`;
+const maxDate = `1-1-${new Date(Date.now()).getFullYear() + 1}`;
+
+const eventsSchema = Joi.object({
+  name: Joi.string().required(),
+
+  startDate: Joi.date().greater(minDate),
+
+  endDate: Joi.date().less(maxDate),
+
+  category: Joi.string()
+    .required()
+    .valid("Session", "OnDayEvent", "Marathon", "Competition"),
+
+  eventDescription: Joi.string().required(),
+
+  eventMobileDescription: Joi.string().required(),
+
+  eventDetails: Joi.string().allow(""),
+
+  eventLocation: Joi.string().required(),
+
+  eventImageID: Joi.string().allow(""),
+});
 
 module.exports = {
   getAllEvents: (req, res) => {
